Limit websocket retries.
Before, the client script would retry for a websocket connection every two seconds.
This caused massive log spam.
Now, it will retry with an exponential timeout, and with a max of 10 retries.
