+++ package.json
@@ -53,8 +53,8 @@
   ],
   "scripts": {
     "prepublish": "webpack ./client/live.js client/live.bundle.js --color --config client/webpack.config.js -p && webpack ./client/index.js client/index.bundle.js --color --config client/webpack.config.js -p",
-    "lint": "eslint bin lib test examples client/{index,live,webpack.config}.js",
-    "beautify-lint": "beautify-lint lib/**.js bin/**.js client/{index,live,webpack.config}.js",
+    "lint": "eslint bin lib test examples client/{index,live,socket,webpack.config}.js",
+    "beautify-lint": "beautify-lint lib/**.js bin/**.js client/{index,live,socket,webpack.config}.js",
     "beautify": "beautify-rewrite lib/**.js bin/**.js client/{index,live,webpack.config}.js",
     "travis": "npm run lint && npm run beautify-lint && node lib/Server.js"
   }
